GenerationService : move to service package
