+++ src/main/java/fr/insee/eno/params/pipeline/PipelineGenerator.java
@@ -2,7 +2,6 @@ package fr.insee.eno.params.pipeline;
 
 import java.util.List;
 
-import fr.insee.eno.GenerationService;
 import fr.insee.eno.generation.Generator;
 import fr.insee.eno.parameters.InFormat;
 import fr.insee.eno.parameters.OutFormat;
@@ -11,6 +10,7 @@ import fr.insee.eno.parameters.PostProcessing;
 import fr.insee.eno.parameters.PreProcessing;
 import fr.insee.eno.postprocessing.Postprocessor;
 import fr.insee.eno.preprocessing.Preprocessor;
+import fr.insee.eno.service.GenerationService;
 
 public interface PipelineGenerator {
