+++ src/test/java/fr/insee/eno/main/DummyTestDDI2JS.java
@@ -4,7 +4,6 @@ import java.io.File;
 
 import org.junit.Test;
 
-import fr.insee.eno.GenerationService;
 import fr.insee.eno.generation.DDI2JSGenerator;
 import fr.insee.eno.postprocessing.Postprocessor;
 import fr.insee.eno.postprocessing.js.JSExternalizeVariablesPostprocessor;
@@ -12,6 +11,7 @@ import fr.insee.eno.postprocessing.js.JSInsertGenericQuestionsPostprocessor;
 import fr.insee.eno.postprocessing.js.JSSortComponentsPostprocessor;
 import fr.insee.eno.postprocessing.js.JSVTLParserPostprocessor;
 import fr.insee.eno.preprocessing.DDIPreprocessor;
+import fr.insee.eno.service.GenerationService;
 
 public class DummyTestDDI2JS {
