+++ src/test/java/fr/insee/eno/main/DummyTestDDI2PDFExamples.java
@@ -21,7 +21,6 @@ import org.apache.fop.apps.FopFactory;
 import org.apache.fop.apps.MimeConstants;
 import org.junit.Test;
 
-import fr.insee.eno.GenerationService;
 import fr.insee.eno.generation.DDI2PDFGenerator;
 import fr.insee.eno.postprocessing.PDFEditStructurePagesPostprocessor;
 import fr.insee.eno.postprocessing.PDFInsertAccompanyingMailsPostprocessor;
@@ -32,6 +31,7 @@ import fr.insee.eno.postprocessing.PDFSpecificTreatmentPostprocessor;
 import fr.insee.eno.postprocessing.PDFTableColumnPostprocessorFake;
 import fr.insee.eno.postprocessing.Postprocessor;
 import fr.insee.eno.preprocessing.DDIPreprocessor;
+import fr.insee.eno.service.GenerationService;
 
 public class DummyTestDDI2PDFExamples {
