+++ src/test/java/fr/insee/eno/main/DummyTestDDI2XForms.java
@@ -4,10 +4,10 @@ import java.io.File;
 
 import org.junit.Test;
 
-import fr.insee.eno.GenerationService;
 import fr.insee.eno.generation.DDI2FRGenerator;
 import fr.insee.eno.postprocessing.NoopPostprocessor;
 import fr.insee.eno.preprocessing.DDIPreprocessor;
+import fr.insee.eno.service.GenerationService;
 
 public class DummyTestDDI2XForms {
