+++ src/test/java/fr/insee/eno/main/DummyTestPoguesXML2DDI.java
@@ -4,10 +4,10 @@ import java.io.File;
 
 import org.junit.Test;
 
-import fr.insee.eno.GenerationService;
 import fr.insee.eno.generation.PoguesXML2DDIGenerator;
 import fr.insee.eno.postprocessing.DDIPostprocessor;
 import fr.insee.eno.preprocessing.PoguesXMLPreprocessorGoToTreatment;
+import fr.insee.eno.service.GenerationService;
 
 public class DummyTestPoguesXML2DDI {
