+++ src/test/java/fr/insee/eno/test/TestDDIToJS.java
@@ -7,7 +7,6 @@ import org.junit.Assert;
 import org.junit.Test;
 import org.xmlunit.diff.Diff;
 
-import fr.insee.eno.GenerationService;
 import fr.insee.eno.generation.DDI2JSGenerator;
 import fr.insee.eno.postprocessing.Postprocessor;
 import fr.insee.eno.postprocessing.js.JSExternalizeVariablesPostprocessor;
@@ -15,6 +14,7 @@ import fr.insee.eno.postprocessing.js.JSInsertGenericQuestionsPostprocessor;
 import fr.insee.eno.postprocessing.js.JSSortComponentsPostprocessor;
 import fr.insee.eno.postprocessing.js.JSVTLParserPostprocessor;
 import fr.insee.eno.preprocessing.DDIPreprocessor;
+import fr.insee.eno.service.GenerationService;
 
 public class TestDDIToJS {
