+++ src/test/java/fr/insee/eno/test/TestDDIToXForm.java
@@ -7,10 +7,10 @@ import org.junit.Assert;
 import org.junit.Test;
 import org.xmlunit.diff.Diff;
 
-import fr.insee.eno.GenerationService;
 import fr.insee.eno.generation.DDI2FRGenerator;
 import fr.insee.eno.postprocessing.fr.FRBrowsingPostprocessor;
 import fr.insee.eno.preprocessing.DDIPreprocessor;
+import fr.insee.eno.service.GenerationService;
 
 public class TestDDIToXForm {
